authManager function implementation added

+++ online_voting_app/src/crypto/cryptoengine.h
@@ -29,7 +29,7 @@ public:
     struct HashResult
     {
         QByteArray hash;
-        int salt;
+        long long salt;
     };
     static CryptoEngine &getInstance();
     CryptoEngine(const CryptoEngine &) = delete;
